Fixed some more reference errors. moved TokStreamer to src/Parsing
