+++ src/Grammar/Program/FunctionGrammar/Assignment.cpp
@@ -1,7 +1,7 @@
 #include "Assignment.h"
 
 Node* Assignment::tryParse(TokStreamer* st){
-	Tok nextTok = st->peekNextToken();
+	Tok nextTok = st->peekNextTok();
 	if(BaseGrammar::expect(nextTok, tokType::KEYWORD)){
 		
 	}
